refactor: edit domain form description
